% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_travel_associated_dengue_cases.R
\docType{data}
\name{cdc_travel_associated_dengue_cases}
\alias{cdc_travel_associated_dengue_cases}
\title{Travel associated dengue cases by year, 2010 - 2023}
\format{
A tibble with 14 rows and 3 variables:
\describe{
\item{Year}{Year}
\item{Travel.status}{Travel status}
\item{Reported.cases}{Number of reported cases}
}
}
\source{
Accessed from \url{https://www.cdc.gov/dengue/data-research/facts-stats/historic-data.html}. Last accessed (Sep 3, 2024)
}
\usage{
cdc_travel_associated_dengue_cases
}
\description{
Travel associated dengue cases by year, 2010 - 2023 in the US, 2010 - 2023 published by the US Centers for Disease Control and Prevention (CDC)
}
\examples{
head(cdc_travel_associated_dengue_cases)
}
\keyword{datasets}
