% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_annual_data.R, R/sl_annual.R
\docType{data}
\name{china_annual_data}
\alias{china_annual_data}
\title{Dengue related data in China}
\format{
A tibble with 16 rows and 5 variables:
\describe{
\item{year}{Year}
\item{dengue.cases.indigenous}{Number of indigenous dengue cases}
\item{dengue.cases.imported}{Number of imported dengue cases}
\item{counties.with.dengue.fever.indigenous}{Number of counties with  dengue fever - indigenous cases}
\item{counties.with.dengue.fever.imported}{Number of counties with  dengue fever - imported cases}
}

A tibble with 16 rows and 5 variables:
\describe{
\item{year}{Year}
\item{dengue.cases}{Number of dengue cases}
\item{epidemic}{Epidemic: 1 if the year is classified as an epidemic year, 0 otherwise}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8997546/table/ijerph-19-03910-t001/?report=objectonly}

World Health Organization. Dengue Fact Sheet and Situation Report, Sri Lanka. 22 July 2022. Available at: \url{https://cdn.who.int/media/docs/default-source/sri-lanka-documents/dengue-fact-sheet_7-2022-srl.pdf?sfvrsn=49021bd_1}. (Accessed: 4 September 2024)
}
\usage{
china_annual_data

china_annual_data
}
\description{
Annual indigenous and imported dengue cases in mainland China, 2005-2020

Annual reported dengue cases in Sri Lanka 1989-2021
}
\examples{
head(china_annual_data)
head(china_annual_data)
}
\keyword{datasets}
