% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_srilanka_wer_to_tidy.R,
%   R/convert_srilankawer_to_tidy.R
\name{convert_srilanka_wer_to_tidy}
\alias{convert_srilanka_wer_to_tidy}
\title{Read weekly epidemiological reports and convert dengue data into tidy format}
\usage{
convert_srilanka_wer_to_tidy(
  year,
  report.url,
  start.date.first,
  end.date.first,
  start.date.last,
  end.date.last,
  week.no
)

convert_srilanka_wer_to_tidy(
  year,
  report.url,
  start.date.first,
  end.date.first,
  start.date.last,
  end.date.last,
  week.no
)
}
\arguments{
\item{year}{year}

\item{report.url}{URL of all WER that you want to extract tables and convert into tidy (The URLs can be extracted from the get_pdflinks_srilanka function).}

\item{start.date.first}{starting date of the first report week (format "2006-12-23")}

\item{end.date.first}{endinging date of the first report week (format "2006-12-23")}

\item{start.date.last}{starting date of the last report week (format "2006-12-23")}

\item{end.date.last}{endinging date of the first report week (format "2006-12-23")}

\item{week.no}{report number vector}

\item{url.part1}{first part of the URL, by default "https://www.epid.gov.lk"}

\item{url.part2}{URL address vector obtained from filter_year_wer_link}
}
\description{
Read weekly epidemiological reports and convert dengue data into tidy format

Read weekly epidemiological reports and convert dengue data into tidy format
}
