% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_year_wer_link.R
\name{filter_year_wer}
\alias{filter_year_wer}
\title{Filter links corresponds to a specific year using the function output get_address}
\usage{
filter_year_wer(year, address)
}
\arguments{
\item{year}{year that you want to extract the reports}

\item{address}{list or the outout from get_addess function}
}
\description{
Filter links corresponds to a specific year using the function output get_address
}
\author{
Thiyanga S Talagala
}
