% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_addresses.R
\name{get_addresses}
\alias{get_addresses}
\title{Extract links of all downloadable files on a webpage}
\usage{
get_addresses(url)
}
\arguments{
\item{url}{url of the webpage that you intend to download files.}
}
\value{
web addresses of all downloadable files
}
\description{
Extract links of all downloadable files on a webpage
}
\author{
Thiyanga S Talagala
}
