% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdflinks_srilanka.R
\name{get_pdflinks_srilanka}
\alias{get_pdflinks_srilanka}
\title{Get URLs of weekly epidemiological reports from Epidemiology Unit, Ministry of Health, Sri Lanka (url https://www.epid.gov.lk/weekly-epidemiological-report/weekly-epidemiological-report)}
\usage{
get_pdflinks_srilanka(
  url =
    "https://www.epid.gov.lk/weekly-epidemiological-report/weekly-epidemiological-report",
  volume.number
)
}
\arguments{
\item{url}{url of the webpage that you intend to download files. The default is URL of the weekly epidemiological reports page}

\item{volume.number}{Volume number of the epidemiological reports (Volumes corresponds to year)}
}
\value{
URLS of PDF files corresponds to the volume number
}
\description{
Get URLs of weekly epidemiological reports from Epidemiology Unit, Ministry of Health, Sri Lanka (url https://www.epid.gov.lk/weekly-epidemiological-report/weekly-epidemiological-report)
}
\author{
Thiyanga S Talagala
}
