% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korea_dengue.R
\docType{data}
\name{korea_dengue}
\alias{korea_dengue}
\title{Imported dengue cases in Korea}
\format{
A tibble with 33 rows and 7 variables:
\describe{
\item{Country}{Country}
\item{Region}{Region}
\item{2011}{Year 2011}
\item{2012}{Year 2012}
\item{2013}{Year 2013}
\item{2014}{Year 2014}
\item{2015}{Year 2015}
}
}
\source{
Je, Sungmo, Wonjun Bae, Jiyeon Kim, Seung Hyeok Seok, and Eung-Soo Hwang. "Epidemiological characteristics and risk factors of dengue infection in Korean travelers." Journal of Korean Medical Science 31, no. 12 (2016): 1863-1873.
}
\usage{
korea_dengue
}
\description{
Data for imported dengue cases were collected from Korea Centers for Disease Control and Prevention (KCDC)
}
\examples{
head(korea_dengue)
}
\keyword{datasets}
