% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max}
\alias{min_max}
\title{Apply min-max transformation}
\usage{
min_max(data, variable.to.minmax, local = FALSE, group.var)
}
\arguments{
\item{data}{tibble that contains cases, groups}

\item{variable.to.minmax}{the variable that we want to tranform using the min-max transformation}

\item{local}{TRUE if you need to apply local minmax transformation}

\item{group.var}{variables that you need to create group-wise}
}
\value{
tibble with minmax transformed applied
}
\description{
Apply min-max transformation
}
