% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philippines_daily_data.R
\docType{data}
\name{philippines_daily_data}
\alias{philippines_daily_data}
\title{Daily number of dengue fever infections in Philippines}
\format{
A tibble with 32701 rows and 5 variables:
\describe{
\item{location}{location}
\item{affected_and_infected}{affected and infected number of cases}
\item{affected_and_killed}{affected and killed number of cases}
\item{date}{date of the week}
\item{region}{region name}
}
}
\source{
\url{https://data.humdata.org/dataset/philippine-dengue-cases-and-deaths?}
}
\usage{
philippines_daily_data
}
\description{
Daily cases of dengue in Philippines
}
\examples{
head(philippines_daily_data)
}
\keyword{datasets}
