% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singapore_weekly_data.R
\docType{data}
\name{singapore_weekly_data}
\alias{singapore_weekly_data}
\title{Weekly number of dengue fever infections in Sri Lanka}
\format{
A tibble with 18772 rows and 6 variables:
\describe{
\item{year}{year}
\item{week}{week number}
\item{cases}{Number of dengue cases}
}
}
\source{
\url{https://www.straitstimes.com/multimedia/graphics/2022/06/singapore-dengue-cases/index.html?shell#:~:text=Singapore\%20is\%20currently\%20seeing\%2030,lower\%20than\%20the\%202020\%20surge}
}
\usage{
singapore_weekly_data
}
\description{
Weekly cases of dengue in Sri Lanka
}
\examples{
head(singapore_weekly_data)
}
\keyword{datasets}
