% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_annual.R
\docType{data}
\name{sl_annual}
\alias{sl_annual}
\title{Annual reported dengue cases in Sri Lanka}
\format{
A tibble with 35 rows and 3 variables:
\describe{
\item{year}{Year}
\item{dengue.cases}{Number of dengue cases}
\item{epidemic}{Epidemic: 1 if the year is classified as an epidemic year, 0 otherwise}
}
}
\source{
Source 1: World Health Organization. Dengue Fact Sheet and Situation Report, Sri Lanka. 22 July 2022. Available at: \url{https://cdn.who.int/media/docs/default-source/sri-lanka-documents/dengue-fact-sheet_7-2022-srl.pdf?sfvrsn=49021bd_1}. (Accessed: 4 September 2024), Source 2: Weekly Dengue Updates, National Dengue Control Unit, Ministry of Health
}
\usage{
sl_annual
}
\description{
Annual reported dengue cases in Sri Lanka 1989-2021
}
\examples{
head(sl_annual)
}
\keyword{datasets}
