% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_dengue_serotype.R
\docType{data}
\name{sl_dengue_serotype}
\alias{sl_dengue_serotype}
\title{Identification of dengue serotypes circulating in Sri Lanka}
\format{
A tibble with 35 rows and 3 variables:
\describe{
\item{year}{Year}
\item{dengue.cases}{Number of dengue cases}
}
}
\source{
Ali, Shahid, Abdul Waheed Khan, Andrew W. Taylor-Robinson, Muhammad Adnan, Shahana Malik, and Saba Gul. "The unprecedented magnitude of the 2017 dengue outbreak in Sri Lanka provides lessons for future mosquito-borne infection control and prevention." Infection, Disease & Health 23, no. 2 (2018): 114-120.
}
\usage{
sl_dengue_serotype
}
\description{
History of dengue dengue serotypes circulating in Sri Lanka.
}
\examples{
head(sl_dengue_serotype)
}
\keyword{datasets}
