% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_province_districts.R
\docType{data}
\name{sl_province_districts}
\alias{sl_province_districts}
\title{Provinces and Districts of Sri Lanka}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 2 columns.
}
\source{
{https://moe.gov.lk/ministry-2/provinces/}
}
\usage{
sl_province_districts
}
\description{
Sri Lanka is divided into 9 provinces, each comprising several districts. Here is a list of the districts and the provinces to which they belong.
\describe{
\item{province}{Name of the province}
\item{district}{Name of the district}
}
}
\examples{
sl_province_districts
}
\keyword{datasets}
