% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_sites.R
\docType{data}
\name{sl_sites}
\alias{sl_sites}
\title{Type and proportion of breeding habitats positive for Aedes aegypti mosquitoes, across provinces in Sri Lanka, 2017.}
\format{
A tibble with 9 rows and 7 variables:
\describe{
\item{Province}{Province}
\item{Discarded_items}{Percentage of water-retaining containers positive for discarded items}
\item{Water_storage_containers_and_tanks}{Percentage of water-retaining containers positive for water storage containers and tanks}
\item{Ponds_and_ornamental_items}{Percentage of water-retaining containers positive for ponds and ornamental items}
\item{Wells_and_tube_wells}{Percentage of water-retaining containers positive for wells and tube wells}
\item{Natural_water_collections}{Percentage of water-retaining containers positive for natural water collections}
\item{Other_miscellaneous_items}{Percentage of water-retaining containers positive for other miscellaneous items}
}
}
\source{
Tissera, H.A., Jayamanne, B.D., Raut, R., Janaki, S.M., Tozan, Y., Samaraweera, P.C., Liyanage, P., Ghouse, A., Rodrigo, C., de Silva, A.M. and Fernando, S.D., 2020. Severe dengue epidemic, Sri Lanka, 2017. Emerging infectious diseases, 26(4), p.682.
head(sl_sites)
}
\usage{
sl_sites
}
\description{
Total number of premises inspected: 279,728; Total number of premises inspected: 279,728; total Ae. aegypti–positive containers: 9,699. Other miscellaneous items include refrigerator trays, nonfunctional cisterns, pet feeding cups, gutters, concrete slabs, and any other water-collecting containers. (Last accessed from the source on 3 September 2024).
}
\keyword{datasets}
