% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srilanka_weekly_data.R
\docType{data}
\name{srilanka_weekly_data}
\alias{srilanka_weekly_data}
\title{Weekly number of dengue fever infections in Sri Lanka}
\format{
A tibble with 23815 rows and 6 variables:
\describe{
\item{year}{year}
\item{week}{week number}
\item{start.date}{starting date of the week}
\item{end.date}{ending date of the week}
\item{district}{district name}
\item{cases}{Number of dengue cases}
}
}
\source{
Weekly Epidemiological Reports, Epidemiology Unit, Ministry of Health, Sri Lanka.
}
\usage{
srilanka_weekly_data
}
\description{
Weekly cases of dengue in Sri Lanka
}
\examples{
head(srilanka_weekly_data)
}
\keyword{datasets}
