% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taiwan_dengue.R
\docType{data}
\name{taiwan_dengue}
\alias{taiwan_dengue}
\title{Indigenous and imported dengue cases in Taiwan, 1987-2023.}
\format{
A tibble with 37 rows and 3 variables:
\describe{
\item{year}{Year}
\item{indigenous.dengue}{Number of indigenous dengue cases}
\item{imported.dengue}{Number of imported dengue}
}
}
\source{
Taiwan Centres for Disease Control,  Available at: \url{https://www.cdc.gov.tw/En/Category/ListContent/bg0g_VU_Ysrgkes_KRUDgQ?uaid=9_Oq7OYHa-l8B05iUwyVvQ}. (Accessed: 4 September 2024), Source 2: Weekly Dengue Updates, National Dengue Control Unit, Ministry of Health
}
\usage{
taiwan_dengue
}
\description{
Annual indigenous and imported dengue cases in Taiwan, 1987-2023.
}
\examples{
head(taiwan_dengue)
}
\keyword{datasets}
