% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.annual.R
\docType{data}
\name{world.annual}
\alias{world.annual}
\title{Annual number of dengue fever infections around the world}
\format{
A tibble with 6750 rows and 4 variables:
\describe{
\item{Entity}{Country or area name}
\item{Code}{Country or area code}
\item{Year}{year}
\item{Incidence}{Number of dengue incidence across all ages}
}
}
\source{
\url{https://ourworldindata.org/grapher/dengue-incidence}
}
\usage{
world.annual
}
\description{
Annual incidence of dengue around the world
}
\examples{
head(world.annual)
}
\keyword{datasets}
