% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_annual_data.R
\docType{data}
\name{china_annual_data}
\alias{china_annual_data}
\title{Dengue related data in China}
\format{
A tibble with 16 rows and 5 variables:
\describe{
\item{year}{Year}
\item{dengue.cases.indigenous}{Number of indigenous dengue cases}
\item{dengue.cases.imported}{Number of imported dengue cases}
\item{counties.with.dengue.fever.indigenous}{Number of counties with  dengue fever - indigenous cases}
\item{counties.with.dengue.fever.imported}{Number of counties with  dengue fever - imported cases}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8997546/table/ijerph-19-03910-t001/?report=objectonly}
}
\usage{
china_annual_data
}
\description{
Annual indigenous and imported dengue cases in mainland China, 2005-2020
}
\examples{
head(china_annual_data)
}
\keyword{datasets}
