% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_srilanka_wer_to_tidy.R
\name{convert_srilanka_wer_to_tidy}
\alias{convert_srilanka_wer_to_tidy}
\title{Read weekly epidemiological reports and convert dengue data into tidy format}
\usage{
convert_srilanka_wer_to_tidy(
  year,
  url.part1 = "https://www.epid.gov.lk",
  url.part2,
  start.date.first,
  end.date.first,
  start.date.last,
  end.date.last,
  week.no
)
}
\arguments{
\item{year}{year}

\item{url.part1}{first part of the URL, by default "https://www.epid.gov.lk"}

\item{url.part2}{URL address vector obtained from filter_year_wer_link}

\item{start.date.first}{starting date of the first report week (format "2006-12-23")}

\item{end.date.first}{endinging date of the first report week (format "2006-12-23")}

\item{start.date.last}{starting date of the last report week (format "2006-12-23")}

\item{end.date.last}{endinging date of the first report week (format "2006-12-23")}

\item{week.no}{report number vector}
}
\description{
Read weekly epidemiological reports and convert dengue data into tidy format
}
